major refactor of control io
