+++ docs/api_reference/trestle.core.control_interface.md
@@ -1,0 +1,2 @@
+::: trestle.core.control_interface
+handler: python
