--- docs/api_reference/trestle.core.control_io.md
@@ -1,2 +1,0 @@
-::: trestle.core.control_io
-handler: python
