+++ docs/api_reference/trestle.core.control_reader.md
@@ -1,0 +1,2 @@
+::: trestle.core.control_reader
+handler: python
