+++ docs/api_reference/trestle.core.control_writer.md
@@ -1,0 +1,2 @@
+::: trestle.core.control_writer
+handler: python
